% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobit1.R
\name{nutrient}
\alias{nutrient}
\title{Nutrient dataset}
\description{
Simulated nutrient concentration data, with 1200 rows and 3 columns.
}
\format{
This data frame contains the following columns:

day - a day of the year

location - one of four locations A B C D

y - a nutrient value, either a number or a censoring threshold
}
