% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobit1.R
\name{tobit1}
\alias{tobit1}
\title{Tobit I family for censored GAM}
\description{
This function implements ...
}
\usage{
tobit1(theta = NULL, link = "identity")
}
\arguments{
\item{theta}{...}

\item{link}{...}
}
\details{
How the model must be specified ...
}
\value{
An object inheriting from class \code{family} ...
}
\examples{
# Generate random data
set.seed(1)
x <- matrix(2*rnorm(300), 100)
yn <- 2*x[,3] + 4*cos(x[,1]*2)
y <- yn + rnorm(100)
ycensored <- pmax(y, 0) # data left-censored at zero

par(mfrow = c(3,3))

# True model
plot(gam(y ~ s(x[,1]) + s(x[,2]) + s(x[, 3])), ylim=c(-5, 5), main = "True")

# Naive estimation
plot(gam(ycensored ~ s(x[,1]) + s(x[,2]) + s(x[, 3])), ylim=c(-5, 5), main = "Naive")

# Tobit I estimation
m <- gam(cbind(ycensored, 0, Inf) ~ s(x[,1]) + s(x[,2]) + s(x[, 3]), family = tobit1)
summary(m)
m$family$thetaest # TO REVISE: THIS IS CALLED SIGMA IN SUMMARY OUTPUT HEADING
plot(m, ylim = c(-5, 5), main = "Tobit I") # WHY ARE CONFIDENCE INTERVALS SO WIDE AT THE EXTREMES?
}
