% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobit2.R
\name{tobit2}
\alias{tobit2}
\title{Tobit II family for censored GAM}
\description{
This function implements ...
}
\usage{
tobit2(theta = NULL, link = "identity")
}
\arguments{
\item{theta}{...}

\item{link}{...}
}
\details{
How the model must be specified ...
}
\value{
An object inheriting from class \code{family} ...
}
\examples{
# Generate random data
set.seed(1)
x <- matrix(2 * rnorm(300), 100)
yn <- 2 * x[, 3] + 4 * cos(x[, 1] * 2)
y <- yn + rnorm(100)
ycensored <- pmax(y, 0) # data left-censored at zero
ycensored2 <- replace(y, yn < 0, 0) # WHAT IS THE DIFFERENCE WITH ycensored?
censid2 <- which(yn < 0)
censid3 <- sample(100, 30)
ycensored3 <- replace(y, censid3, 0)

# Fitting alternative models
m1 <- gam(cbind(ycensored, 0, Inf) ~ s(x[, 1]) + s(x[, 2]) + s(x[, 3]), family = tobit2)
plot(m1, ylim = c(-5, 5))

m2 <- gam(cbind(ycensored2, replace(rep(-Inf, 100), censid2, 0), Inf) ~ s(x[, 1]) + s(x[, 2]) + s(x[, 3]), family = tobit2)
plot(m2, ylim = c(-5, 5))

m3 <- gam(cbind(ycensored3, replace(rep(0, 100), censid3,-1)) ~ s(x[, 1]) + s(x[, 2]) + s(x[, 3]), family = tobit2)
plot(m3, ylim = c(-5, 5))

}
